docs(libs): updating doc-blocks in libs
