+++ libs/hub/src/index.ts
@@ -1,2 +1,4 @@
 export * from './lib/hub.module';
+
 export { HubService } from './lib/hub.service';
+export { HubServiceActionInterface } from './lib/hub.typing';
