+++ libs/hub/src/lib/hub.typing.ts
@@ -1,0 +1,8 @@
+export type fn = (params: any) => any;
+
+export interface HubServiceActionInterface {
+  event: string;
+  data: any;
+  callback: fn;
+  action: 'emit' | 'on' | 'off';
+}
