+++ libs/jwt/src/index.ts
@@ -1,1 +1,3 @@
 export * from './lib/jwt.module';
+
+export { JWTService } from './lib/jwt.service';
