+++ libs/responsive/src/index.ts
@@ -1,5 +1,5 @@
 export * from './lib/responsive.module';
-export {
-  RESPONSIVE_BREAKPOINTS_TOKEN,
-  ResponsiveBreakpointDirective
-} from './lib/responsive.directive';
+
+export { Breakpoints } from './lib/responsive.typing';
+export { RESPONSIVE_BREAKPOINTS_TOKEN } from './lib/responsive.token';
+export { ResponsiveBreakpointDirective } from './lib/responsive.directive';
