+++ libs/responsive/src/lib/responsive.module.ts
@@ -2,11 +2,8 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { SupportModule } from '@ng-lab/support';
-import {
-  RESPONSIVE_BREAKPOINTS_TOKEN,
-  DEFAULT_RESPONSIVE_BREAKPOINTS_TOKEN,
-  ResponsiveBreakpointDirective
-} from './responsive.directive';
+import { ResponsiveBreakpointDirective } from './responsive.directive';
+import { RESPONSIVE_BREAKPOINTS_TOKEN, DEFAULT_RESPONSIVE_BREAKPOINTS_TOKEN } from './responsive.token';
 
 @NgModule({
   imports: [CommonModule, SupportModule],
