+++ libs/responsive/src/lib/responsive.typing.ts
@@ -1,0 +1,7 @@
+export interface Breakpoints {
+  XS: number;
+  SM: number;
+  MD: number;
+  LG: number;
+  XL: number;
+}
